+++ apps/judicial-system/api/src/app/modules/indictment-count/index.ts
@@ -2,4 +2,4 @@ export { IndictmentCount } from './models/indictmentCount.model'
 export { CreateIndictmentCountInput } from './dto/createIndictmentCount.input'
 export { UpdateIndictmentCountInput } from './dto/updateIndictmentCount.input'
 export { DeleteIndictmentCountInput } from './dto/deleteIndictmentCount.input'
-export { DeleteIndictmentCountResponse } from './models/delete.response'
+export { DeleteIndictmentCountResponse } from './models/deleteIndictmentCount.response'
