+++ apps/judicial-system/api/src/app/modules/indictment-count/models/deleteIndictmentCount.response.ts
